Archive autotools-based QuantumESPRESSO easyblock and switch default to CMakeMake-based easyblock
